+++ src/actions/youtube_video_actions/channel_actions.js
@@ -2,7 +2,7 @@ import * as YoutubeVideoAPI from 'util/youtube_video_util/index.js';
 
 export const RECEIVE_CHANNEL_INFO = 'RECEIVE_CHANNEL_INFO';
 export const RECEIVE_CHANNEL_VIDEOS = 'RECEIVE_CHANNEL_VIDEOS';
-export const RECEIVE_CHANNEL_DETAIL = 'RECEIVE_CHANNEL_DETAIL';
+// export const RECEIVE_CHANNEL_DETAIL = 'RECEIVE_CHANNEL_DETAIL';
 
 export const receiveChannelInfo = channel => ({
   type: RECEIVE_CHANNEL_INFO,
@@ -14,20 +14,20 @@ export const receiveChannelVideos = videos => ({
   videos
 });
 
-export const receiveChannelDetail = detail => ({
-  type: RECEIVE_CHANNEL_DETAIL,
-  detail
-});
-
-export const fetchChannelDetails = channelId => dispatch => {
-  return YoutubeVideoAPI.fetchChannelDetails(channelId).then(
-    res => res.json()
-  ).then(
-    detail => dispatch(receiveChannelDetail(detail))
-  ).catch(
-    err => console.log(err)
-  );
-};
+// export const receiveChannelDetail = detail => ({
+//   type: RECEIVE_CHANNEL_DETAIL,
+//   detail
+// });
+//
+// export const fetchChannelDetails = channelId => dispatch => {
+//   return YoutubeVideoAPI.fetchChannelDetails(channelId).then(
+//     res => res.json()
+//   ).then(
+//     detail => dispatch(receiveChannelDetail(detail))
+//   ).catch(
+//     err => console.log(err)
+//   );
+// };
 
 export const fetchChannelInfo = channelId => dispatch => {
   return YoutubeVideoAPI.fetchChannelInfo(channelId).then(
