+++ src/reducers/root_reducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import channelsReducer from './channels_reducer';
-import channelDetailsReducer from './channel_details_reducer';
+// import channelDetailsReducer from './channel_details_reducer';
 import queryReducer from './query_reducer';
 import recommendedReducer from './recommended_reducer';
 import searchResultReducer from './search_result_reducer';
@@ -13,7 +13,7 @@ import playerDetailsReducer from './player_details_reducer';
 
 const rootReducer = combineReducers({
   channels: channelsReducer,
-  channelDetails: channelDetailsReducer,
+  // channelDetails: channelDetailsReducer,
   query: queryReducer,
   recommended: recommendedReducer,
   searchResult: searchResultReducer,
